Add support for images in game
The README.md file contains source info for free image use.
The index.html file has a "defer" added to the script so the JS will
be executed after the HTML page loads.
The rps-game.js file now lays out the page with flex boxes and images.
Changes will be made to the layout to add more style.