Display results of last round
Previously, the results of the final round were not being shown,
because everything was deleted and a final result message and new
game button were put in their place. Now the final results message
is placed in the scorebar area, the final round results are shown as
usual, and the rock/paper/scissors buttons are replaced with the
new game button.